+++ types/surah.d.ts
@@ -2,5 +2,8 @@ export type Surah = {
   number: number;
   name: string;
   englishName: string;
+  englishNameTranslation?: string;
+  numberOfAyahs: number;
+  revelationType?: string;
   startingPage: number | undefined;
 };